// Segundo parcial/Clases/13CSR/CrudProducto/src/interfaces/IProducto.ts
// APLICACIONES WEB 2
// Nombres y Apellidos: Winter Aníbal Meza Jiménez.
// Curso: Sexto "B" 2022(1).
// Clase: Client Side Render CRUD.
// Fecha: 30/6/2022- Modificado: Miércoles 13/7/2022.
// Docente: Ing. Jhon Antonio Cevallos Macías, Mg.

// Generated by https://quicktype.io
// Se exportan la interfaz IResProducto con sus respectivos atributos definidos.
export interface IResProducto {
    total:     number;
    productos: Producto[];
}
// Se exportan la interfaz Producto con sus respectivos atributos definidos.
export interface Producto {
    _id?:    string;
    nombre: string;
    estado?: boolean;
    precio: number;
    costo:  number;
    minimo: number;
    stock:  number;
    __v?:    number;
}